refactor: split src/index.ts into per-responsibility modules
Pure code move with no behavior changes:
- src/cardvault/types.ts: CardVault API (snake_case) types
- src/cardvault/client.ts: CardVault request helpers, error logging,
  and shared constants
- src/logger.ts: shared pino logger
- src/transform.ts: response transformation utilities
- src/schemas.ts: zod output schemas and derived public types
- src/tools/: one file per tool (search, prints, card-detail, products)
- src/server.ts: createServer() factory
- src/index.ts: HTTP routing and the default export only
